moved the xbox controller class into the components folder
