+++ src/main/java/frc/robot/components/DTXboxController.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.components;
 
 import java.util.ArrayList;
 import com.ctre.phoenix.time.StopWatch;
